+++ src/inputs.ts
@@ -4,6 +4,7 @@ import * as core from "@actions/core";
 const INPUT_NAME_CONFIG_PATH = "configuration-path";
 const INPUT_NAME_DRY_RUN = "dry-run";
 const INPUT_NAME_REPO_TOKEN = "repo-token";
+const INPUT_NAME_SVGO_OPTIONS = "svgo-options";
 
 const FALSE = "false";
 const TRUE = "true";
@@ -11,6 +12,7 @@ const TRUE = "true";
 
 export type RawActionConfig = {
   readonly "dry-run"?: string;
+  readonly "svgo-options"?: string;
 }
 
 
@@ -20,6 +22,11 @@ export class ActionConfig {
 
   constructor(config?: RawActionConfig) {
     this.config = config || { };
+  }
+
+  public getSvgoOptionsPath(): string {
+    return this.config["svgo-options"]
+      || core.getInput(INPUT_NAME_SVGO_OPTIONS, { required: false });
   }
 
   public isDryRun(): boolean {