+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Versioning].
 
 ## [Unreleased]
 
+- Configure the SVGO options file. ([#99])
 - Fix `required` value for "configuration-path" input. ([#100])
 
 ## [0.2.1] - 2020-03-02
@@ -53,4 +54,5 @@ Versioning].
 [#91]: https://github.com/ericcornelissen/svgo-action/pull/91
 [#92]: https://github.com/ericcornelissen/svgo-action/pull/92
 [#97]: https://github.com/ericcornelissen/svgo-action/pull/97
+[#99]: https://github.com/ericcornelissen/svgo-action/pull/99
 [#100]: https://github.com/ericcornelissen/svgo-action/pull/100